Add translations, move some things around
I think all things (currently, though in Sage 6.5 more will be there) in the doc are now listed as well.  Fixed a link - well, that is, William made most sage.math/home/user directories not readable, so I found another copy online.

I also tried to move things out of the standard documentation and into some more categories.  Really these should not be bullets but different header levels, but I stuck with what was there for now in the hopes it will be easy to review.