Update package info in components.yaml
Fix stale URLs for a number of packages in the components.yaml file and
remove two packages from the list, Jinja and weave.  Both of these
packages are outdated components (replaced respectively by Jinja2 and
Cython) and are no longer included in Sagemath distributions.

Additionally updated formatting in links-components.html to handle
packages for which no appropriate URL is available, such that the url
field in components.yaml is the empty string.